requirements.txt
aiohappyeyeballs>=2.4.4
aiohttp>=3.11.11
aiosignal>=1.3.2
altair>=5.5.0
annotated-types>=0.7.0
anyio>=4.7.0
async-timeout>=4.0.3
attrs>=24.3.0
beautifulsoup4>=4.12.3
blinker>=1.9.0
cachetools>=5.5.0
certifi>=2024.12.14
charset-normalizer>=3.4.1
click>=8.1.8

colorama>=0.4.6
contourpy>=1.3.1
cycler>=0.12.1
dataclasses-json>=0.6.7
datasets>=3.2.0
dill>=0.3.8
exceptiongroup>=1.2.2
faiss-cpu>=1.8.0
filelock>=3.16.1
fonttools>=4.55.3
frozenlist>=1.5.0
fsspec>=2024.9.0
ftfy>=6.3.1
future>=1.0.0
gitdb>=4.0.11
GitPython>=3.1.43
greenlet>=3.1.1
h11>=0.14.0
httpcore>=1.0.7
httpx>=0.28.1
httpx-sse>=0.4.0
huggingface-hub>=0.27.0
idna>=3.10
Jinja2>=3.1.5
joblib>=1.4.2
jsonpatch>=1.33
jsonpointer>=3.0.0
jsonschema>=4.23.0
jsonschema-specifications>=2024.10.1
kiwisolver>=1.4.8
langchain>=0.3.13
langchain-community>=0.3.13
langchain-core>=0.3.28
langchain-experimental>=0.3.4
langchain-text-splitters>=0.3.4
langsmith>=0.1.147
llvmlite>=0.43.0
lxml>=5.1.0
markdown-it-py>=3.0.0
MarkupSafe>=3.0.2
marshmallow>=3.23.2
matplotlib>=3.10.0
mdurl>=0.1.2
more-itertools>=10.5.0
mpmath>=1.3.0
multidict>=6.1.0
multiprocess>=0.70.16
mypy-extensions>=1.0.0
narwhals>=1.19.1
networkx>=3.4.2
numba>=0.60.0
numpy>=1.26.4
open_clip_torch>=2.29.0
openai-whisper>=20240930
orjson>=3.10.12
packaging>=24.2
pandas>=2.2.3
pillow>=11.0.0
propcache>=0.2.1
protobuf>=5.29.2
pyarrow>=18.1.0
pydantic>=2.10.4
pydantic-settings>=2.7.0
pydantic_core>=2.27.2
pydeck>=0.9.1
Pygments>=2.18.0
pyparsing>=3.2.0
PyPDF2>=3.0.1
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
pytz>=2024.2
PyYAML>=6.0.2
referencing>=0.35.1
regex>=2024.11.6
requests>=2.32.3
requests-toolbelt>=1.0.0
rich>=13.9.4
rpds-py>=0.22.3
safetensors>=0.4.5
scikit-learn>=1.6.0
scipy>=1.14.1
sentence-transformers>=3.3.1
six>=1.17.0
smmap>=5.0.1
sniffio>=1.3.1
soupsieve>=2.6
SQLAlchemy>=2.0.36
streamlit>=1.43.0
streamlit-option-menu>=0.4.0
sympy>=1.13.1
tenacity>=8.5.0
threadpoolctl>=3.5.0
tiktoken>=0.8.0
timm>=1.0.12
tokenizers>=0.21.0
toml>=0.10.2
torch>=2.5.1
torchvision>=0.20.1
tornado>=6.4.2
tqdm>=4.67.1
transformers>=4.47.1
typing-inspect>=0.9.0
typing_extensions>=4.12.2
tzdata>=2024.2
urllib3>=2.3.0
watchdog>=6.0.0
wcwidth>=0.2.13
xxhash>=3.5.0
yarl>=1.18.3
google-generativeai

git+https://github.com/openai/CLIP.git
groq>=1.0.0
reportlab>=4.0
openpyxl>=3.1.0
